reference-app/trial/requirements.txt
Flask
Flask-Opentracing
requests
pylint
gunicorn
jaeger-client
prometheus-client
opentelemetry-api
opentelemetry-instrumentation
opentelemetry-distro
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-sdk
prometheus-flask-exporter